# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Source Accounting API",
    "description": "Free general chat and paid annual Agent workflows. All research outputs are drafts.",
    "version": "0.7.0"
  },
  "paths": {
...
    "/api/v1/me/preferences": {
      "put": {
        "tags": [
          "account"
        ],
        "summary": "Preferences",
        "operationId": "preferences_api_v1_me_preferences_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preferences"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/feedback": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "Feedback",
        "operationId": "feedback_api_v1_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "Read Notification",
        "operationId": "read_notification_api_v1_notifications__notification_id__read_post",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/library": {
      "get": {
        "tags": [
          "open library"
        ],
        "summary": "Library",
        "operationId": "library_api_v1_library_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "topic",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 60,
              "default": "",
              "title": "Topic"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 40,
              "default": "",
              "title": "Kind"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 24,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/library/{item_id}": {
      "get": {
        "tags": [
          "open library"
        ],
        "summary": "Library Item",
        "operationId": "library_item_api_v1_library__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/editorial/sources/{source_id}/review": {
      "post": {
        "tags": [
          "open library"
        ],
        "summary": "Editorial Review",
        "operationId": "editorial_review_api_v1_editorial_sources__source_id__review_post",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditorialDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EditorialDecision": {
        "properties": {
          "expected_policy_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Policy Version"
          },
          "content_sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Content Sha256"
          },
          "decision": {
            "type": "string",
            "enum": [
              "approved",
              "changes_requested"
            ],
            "title": "Decision"
          },
          "review_note": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 20,
            "title": "Review Note"
          },
          "checked_reference_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Checked Reference Ids"
          },
          "confirm_actual_review_performed": {
            "type": "boolean",
            "const": true,
            "title": "Confirm Actual Review Performed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_policy_version",
          "content_sha256",
          "decision",
          "review_note",
          "checked_reference_ids",
          "confirm_actual_review_performed"
        ],
        "title": "EditorialDecision"
      },
...
      "FeedbackCreate": {
        "properties": {
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "category": {
            "type": "string",
            "enum": [
              "accuracy",
              "citation",
              "privacy",
              "copyright",
              "interface",
              "other"
            ],
            "title": "Category"
          },
          "message": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 10,
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "category",
          "message"
        ],
        "title": "FeedbackCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Preferences": {
        "properties": {
          "reduced_motion": {
            "type": "boolean",
            "title": "Reduced Motion",
            "default": false
          },
          "compact": {
            "type": "boolean",
            "title": "Compact",
            "default": false
          },
          "default_framework": {
            "type": "string",
            "enum": [
              "US_GAAP",
              "IFRS",
              "UNKNOWN"
            ],
            "title": "Default Framework",
            "default": "US_GAAP"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Preferences"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}